+++ back-end/lib/app.js
@@ -12,9 +12,8 @@ const authenticate = authenticator({
 app.use(require("body-parser").json());
 
 /* WORKING CORS (including preflight) */
-app.use(cors({credentials: true, origin: true}))
-app.options('*', cors({credentials: true, origin: true})) // include before other routes
-
+app.use(cors({ credentials: true, origin: true }));
+app.options("*", cors({ credentials: true, origin: true })); // include before other routes
 
 app.get("/", authenticate, (req, res) => {
   res.send(["<h1>ECE DevOps Chat</h1>"].join(""));
@@ -33,28 +32,50 @@ app.get("/", authenticate, (req, res) => {
 
 /* get channels => db list channels */
 app.get("/channels", authenticate, async (req, res) => {
-  const channels = await db.channels.list();
-  // TODO faire le tri entre les channels que l'user a le droit de voir
-  res.json(channels);
+  let channels = await db.channels.list(); // get every channel from DB
+  const reqEmail = req.headers["email"]; // get email of user asking for channels
+  // filter channels on user is the owner or one of the members
+  channels = channels.filter(
+    (channel) =>
+      channel.owner === reqEmail || channel.members.includes(reqEmail)
+  );
+  // sending back the filtered channels
+  if (!Array.isArray(channels) || !channels.length) {
+    // array does not exist, is not an array, or is empty
+    res.status(204); // No Content
+  } else {
+    res.status(200).json(channels); // OK
+  }
 });
 
 /* post channels => db create channel */
 app.post("/channels", authenticate, async (req, res) => {
   const channel = await db.channels.create(req.body);
-  res.status(201).json(channel);
+  res.status(201).json(channel); // Created
 });
 
 /* get channels:id => db get channel */
 app.get("/channels/:id", authenticate, async (req, res) => {
   const channel = await db.channels.get(req.params.id);
-  // TODO check si l'user a le droit de voir cette channel
-  res.json(channel);
+  // check si l'user a le droit de voir cette channel
+  const reqEmail = req.headers["email"];
+  if (channel.owner === reqEmail || channel.members.includes(reqEmail)) {
+    res.status(200).json(channel); // OK
+  } else {
+    res.status(403); // Forbidden
+  }
 });
 
 /* put channels:id => db update channel */
 app.put("/channels/:id", authenticate, async (req, res) => {
-  const channel = await db.channels.update(req.body);
-  res.json(channel);
+  let channel = await db.channels.get(req.params.id);
+  const reqEmail = req.headers["email"];
+  if (channel.owner === reqEmail || channel.members.includes(reqEmail)) {
+    channel = await db.channels.update(req.body);
+    res.status(201).json(channel); // Created
+  } else {
+    res.status(403); // Forbidden
+  }
 });
 
 //////////////
@@ -62,13 +83,25 @@ app.put("/channels/:id", authenticate, async (req, res) => {
 //////////////
 
 app.get("/channels/:id/messages", authenticate, async (req, res) => {
-  const messages = await db.messages.list(req.params.id);
-  res.json(messages);
+  const channel = await db.channels.get(req.params.id);
+  const reqEmail = req.headers["email"];
+  if (channel.owner === reqEmail || channel.members.includes(reqEmail)) {
+    const messages = await db.messages.list(req.params.id);
+    res.status(200).json(messages); // OK
+  } else {
+    res.status(403); // Forbidden
+  }
 });
 
 app.post("/channels/:id/messages", authenticate, async (req, res) => {
-  const message = await db.messages.create(req.params.id, req.body);
-  res.status(201).json(message);
+  const channel = await db.channels.get(req.params.id);
+  const reqEmail = req.headers["email"];
+  if (channel.owner === reqEmail || channel.members.includes(reqEmail)) {
+    const message = await db.messages.create(req.params.id, req.body);
+    res.status(201).json(message); // Created
+  } else {
+    res.status(403); // Forbidden
+  }
 });
 
 ///////////